+++ android/filament-android/src/main/java/com/google/android/filament/Texture.java
@@ -507,6 +507,23 @@ public class Texture {
     }
 
     /**
+     * Short-lived object similar to PixelBufferDescriptor.
+     */
+    public static class SynchronizedImage {
+        public Long nativeEglImage;
+
+        @Nullable public Object handler;
+        @Nullable public Runnable callback;
+
+        public SynchronizedImage(@NonNull Long nativeEglImage,
+                @Nullable Object handler, @Nullable Runnable callback) {
+            this.nativeEglImage = nativeEglImage;
+            this.handler = handler;
+            this.callback = callback;
+        }
+    }
+
+    /**
      * Options of {@link #generatePrefilterMipmap}
      */
     public static class PrefilterOptions {
@@ -894,6 +911,11 @@ public class Texture {
         nSetExternalImage(getNativeObject(), engine.getNativeObject(), eglImage);
     }
 
+    public void setSynchronizedImage(@NonNull Engine engine, SynchronizedImage image) {
+        nSetSynchronizedImage(getNativeObject(), engine.getNativeObject(),
+                image.nativeEglImage, image.handler, image.callback);
+    }
+
     /**
      * Specifies the external stream to associate with this <code>Texture</code>.
      *
@@ -1075,6 +1097,9 @@ public class Texture {
 
     private static native void nSetExternalImage(
             long nativeObject, long nativeEngine, long eglImage);
+
+    private static native void nSetSynchronizedImage(long nativeObject, long nativeEngine,
+            long eglImage, Object handler, Runnable callback);
 
     private static native void nSetExternalStream(long nativeTexture,
             long nativeEngine, long nativeStream);